requirements.txt
fastapi>=0.104.0
uvicorn>=0.24.0
bcrypt>=4.1.0
slowapi>=0.1.9
pydantic[email]>=2.0.0
pydantic-settings>=2.1.0
pyyaml>=6.0.1
google-adk>=0.3.0
litellm>=1.0.0
python-multipart>=0.0.6
matplotlib>=3.8.0
numpy>=1.26.0
